Create sublist.go
Hello,
I do not know if you have tried to partition the subscriptions by the length of the tokens. I have been testing and the results are better and the changes are insignificant. I only focus on the changes in subscription and comparison, although in the management of the cache also improves